fix style

+++ DataStructure/SegmentTree/with STL/lowerBoundInRange.cpp
@@ -14,22 +14,22 @@ Long combine(Long x, Long y){
 
 struct SegmentTree {
 	vector<Long> t[2 * MX]; //O(nlogn)
-	Long n;
+	int n;
 	
-	void clear(Long n) { 
-		for (Long i = 0; i < 2 * n; i++) {
+	void clear(int n) { 
+		for (int i = 0; i < 2 * n; i++) {
 			t[i].clear();
 		}
 		this->n = n;
 	}
 
-	void build(vector<Long> &a, Long id, Long tl, Long tr) { //O(nlogn)
+	void build(vector<Long> &a, int id, int tl, int tr) { //O(nlogn)
 		if (tl == tr) {
 			t[id] = {a[tl]};
 		} else {
-			Long tm = (tl + tr) / 2;
-			Long left = id + 1;
-			Long right = id + 2 * (tm - tl + 1) ;
+			int tm = (tl + tr) / 2;
+			int left = id + 1;
+			int right = id + 2 * (tm - tl + 1) ;
 			build(a, left , tl, tm);
 			build(a, right, tm + 1, tr);
 			merge(t[left].begin(), t[left].end(), t[right].begin(), t[right].end(), back_inserter(t[id]));
@@ -41,7 +41,7 @@ struct SegmentTree {
 		build(a , 1 , 0 , n - 1);
 	}
 
-	Long query(Long l , Long r, Long x, Long id, Long tl, Long tr) { //O(log^2 n)
+	Long query(int l, int r, Long x, int id, int tl, int tr) { //O(log^2 n)
 		//find the smallest number greater or equal to x
 		if (tr < l || tl > r) return INF;
 		if (l <= tl && tr <= r) {
@@ -49,13 +49,13 @@ struct SegmentTree {
 			if (it != t[id].end()) return *it;
 			else return INF;
 		}
-		Long tm = (tl + tr) / 2;
-		Long left = id + 1;
-		Long right = id + 2 * (tm - tl + 1) ;
+		int tm = (tl + tr) / 2;
+		int left = id + 1;
+		int right = id + 2 * (tm - tl + 1) ;
 		return combine(query(l, r, x , left, tl, tm) , query(l, r, x , right, tm + 1, tr));
 	}
 	
-	Long query(Long l , Long r, Long x) {
+	Long query(int l, int r, Long x) {
 		assert(n > 0);
 		return query(l , r , x , 1 , 0 , n - 1);
 	}
